@@ -3,84 +3,71 @@ package main
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
-	"sort"
-	"strings"
-	"time"
-
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/log/level"
-	"github.com/oklog/ulid"
-	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/prometheus/pkg/labels"
+	"github.com/go-kit/log"
+	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/promql/parser"
+	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
 	"github.com/prometheus/prometheus/tsdb/chunks"
 	tsdb_errors "github.com/prometheus/prometheus/tsdb/errors"
-	"github.com/thanos-io/thanos/pkg/block"
+	"github.com/prometheus/prometheus/tsdb/index"
+	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
-	"github.com/thanos-io/thanos/pkg/objstore/client"
-	"github.com/thanos-io/thanos/pkg/runutil"
+	"golang.org/x/exp/slices"
+	"path/filepath"
+	"strings"
+	"time"
 )
 
-var merr tsdb_errors.MultiError
-
-func analyze(objStoreConfig []byte, id *string, dir *string, limit *int, logger log.Logger, metrics *prometheus.Registry) error {
-	bkt, err := client.NewBucket(logger, objStoreConfig, metrics, "bucket")
-	if err != nil {
+func analyze(bkt objstore.Bucket, id, dir *string, analyzeLimit *int, analyzeMatchers *string, logger log.Logger) error {
+	ctx := context.Background()
+	if err := downloadBlock(ctx, *dir, *id, bkt, logger); err != nil {
 		return err
 	}
-
-	// Ensure we close up everything properly.
-	defer func() {
-		if err != nil {
-			runutil.CloseWithLogOnErr(logger, bkt, "bucket client")
-		}
-	}()
-
-	ctx, _ := context.WithCancel(context.Background())
-	bdir := filepath.Join(*dir, *id)
-	begin := time.Now()
-	err = block.Download(ctx, logger, bkt, ulid.MustParse(*id), bdir)
-	if err != nil {
-		return errors.Wrapf(err, "download block %s", *id)
-	}
-	level.Info(logger).Log("msg", "downloaded block", "id", *id, "duration", time.Since(begin))
-
-	return analyzeBlock(*dir, *id, *limit)
+	return analyzeBlock(*dir, *id, *analyzeLimit, *analyzeMatchers)
 }
 
-// https://github.com/prometheus/prometheus/blob/6573bf42f2431470e375faa515f282eb36865007/cmd/promtool/tsdb.go#L416
-func analyzeBlock(path, blockID string, limit int) error {
+// https://github.com/prometheus/prometheus/blob/main/cmd/promtool/tsdb.go#L415
+func analyzeBlock(path, blockID string, limit int, matchers string) error {
+	var (
+		selectors []*labels.Matcher
+		err       error
+	)
+	if len(matchers) > 0 {
+		selectors, err = parser.ParseMetricSelector(matchers)
+		if err != nil {
+			return err
+		}
+	}
 	db, block, err := openBlock(path, blockID)
 	if err != nil {
 		return err
 	}
 	defer func() {
-		merr.Add(err)
-		merr.Add(db.Close())
-		err = merr.Err()
+		err = tsdb_errors.NewMulti(err, db.Close()).Err()
 	}()
 
 	meta := block.Meta()
 	fmt.Printf("Block ID: %s\n", meta.ULID)
-	m, err := metadata.Read(filepath.Join(path, blockID))
+	m, err := metadata.ReadFromDir(filepath.Join(path, blockID))
 	if err != nil {
-		return errors.Wrapf(err, "fail to read meta.json for %s", blockID)
+		return fmt.Errorf("fail to read meta.json for %s: %w", blockID, err)
 	}
 	fmt.Printf("Thanos Labels: %s\n", labelsToString(m.Thanos.Labels))
 
 	// Presume 1ms resolution that Prometheus uses.
 	fmt.Printf("Duration: %s\n", (time.Duration(meta.MaxTime-meta.MinTime) * 1e6).String())
-	fmt.Printf("Series: %d\n", meta.Stats.NumSeries)
+	fmt.Printf("Total Series: %d\n", meta.Stats.NumSeries)
+	if len(matchers) > 0 {
+		fmt.Printf("Matcher: %s\n", matchers)
+	}
 	ir, err := block.Index()
 	if err != nil {
 		return err
 	}
 	defer ir.Close()
 
-	allLabelNames, err := ir.LabelNames()
+	allLabelNames, err := ir.LabelNames(selectors...)
 	if err != nil {
 		return err
 	}
@@ -93,7 +80,7 @@ func analyzeBlock(path, blockID string, limit int) error {
 	postingInfos := []postingInfo{}
 
 	printInfo := func(postingInfos []postingInfo) {
-		sort.Slice(postingInfos, func(i, j int) bool { return postingInfos[i].metric > postingInfos[j].metric })
+		slices.SortFunc(postingInfos, func(a, b postingInfo) bool { return a.metric > b.metric })
 
 		for i, pc := range postingInfos {
 			if i >= limit {
@@ -107,24 +94,59 @@ func analyzeBlock(path, blockID string, limit int) error {
 	labelpairsUncovered := map[string]uint64{}
 	labelpairsCount := map[string]uint64{}
 	entries := 0
-	p, err := ir.Postings("", "") // The special all key.
-	if err != nil {
-		return err
+	var (
+		p    index.Postings
+		refs []storage.SeriesRef
+	)
+	if len(matchers) > 0 {
+		p, err = tsdb.PostingsForMatchers(ir, selectors...)
+		if err != nil {
+			return err
+		}
+		// Expand refs first and cache in memory.
+		// So later we don't have to expand again.
+		refs, err = index.ExpandPostings(p)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Matched series: %d\n", len(refs))
+		p = index.NewListPostings(refs)
+	} else {
+		p, err = ir.Postings("", "") // The special all key.
+		if err != nil {
+			return err
+		}
 	}
-	lbls := labels.Labels{}
+
 	chks := []chunks.Meta{}
+	builder := labels.ScratchBuilder{}
+	first := true
+	splitLabels := make(map[string]bool) // labels appearing in all series
 	for p.Next() {
-		if err = ir.Series(p.At(), &lbls, &chks); err != nil {
+		if err = ir.Series(p.At(), &builder, &chks); err != nil {
 			return err
 		}
 		// Amount of the block time range not covered by this series.
 		uncovered := uint64(meta.MaxTime-meta.MinTime) - uint64(chks[len(chks)-1].MaxTime-chks[0].MinTime)
-		for _, lbl := range lbls {
+		builder.Labels().Range(func(lbl labels.Label) {
 			key := lbl.Name + "=" + lbl.Value
 			labelsUncovered[lbl.Name] += uncovered
 			labelpairsUncovered[key] += uncovered
 			labelpairsCount[key]++
 			entries++
+		})
+		if first {
+			for _, l := range builder.Labels() {
+				if l.Name != "__name__" {
+					splitLabels[l.Name] = true
+				}
+			}
+			first = false
+		}
+		for l, _ := range splitLabels {
+			if !builder.Labels().Has(l) {
+				delete(splitLabels, l)
+			}
 		}
 	}
 	if p.Err() != nil {
@@ -132,6 +154,11 @@ func analyzeBlock(path, blockID string, limit int) error {
 	}
 	fmt.Printf("Postings (unique label pairs): %d\n", len(labelpairsUncovered))
 	fmt.Printf("Postings entries (total label pairs): %d\n", entries)
+	tmp := []string{}
+	for k, _ := range splitLabels {
+		tmp = append(tmp, k)
+	}
+	fmt.Printf("Label names appearing in all Series: [%s]\n", strings.Join(tmp, ", "))
 
 	postingInfos = postingInfos[:0]
 	for k, m := range labelpairsUncovered {
@@ -159,7 +186,7 @@ func analyzeBlock(path, blockID string, limit int) error {
 
 	postingInfos = postingInfos[:0]
 	for _, n := range allLabelNames {
-		values, err := ir.SortedLabelValues(n)
+		values, err := ir.SortedLabelValues(n, selectors...)
 		if err != nil {
 			return err
 		}
@@ -175,7 +202,7 @@ func analyzeBlock(path, blockID string, limit int) error {
 
 	postingInfos = postingInfos[:0]
 	for _, n := range allLabelNames {
-		lv, err := ir.SortedLabelValues(n)
+		lv, err := ir.SortedLabelValues(n, selectors...)
 		if err != nil {
 			return err
 		}
@@ -185,7 +212,7 @@ func analyzeBlock(path, blockID string, limit int) error {
 	printInfo(postingInfos)
 
 	postingInfos = postingInfos[:0]
-	lv, err := ir.SortedLabelValues("__name__")
+	lv, err := ir.SortedLabelValues("__name__", selectors...)
 	if err != nil {
 		return err
 	}
@@ -194,6 +221,7 @@ func analyzeBlock(path, blockID string, limit int) error {
 		if err != nil {
 			return err
 		}
+		postings = index.Intersect(postings, index.NewListPostings(refs))
 		count := 0
 		for postings.Next() {
 			count++
@@ -205,7 +233,8 @@ func analyzeBlock(path, blockID string, limit int) error {
 	}
 	fmt.Printf("\nHighest cardinality metric names:\n")
 	printInfo(postingInfos)
-	return os.RemoveAll(filepath.Join(path, blockID))
+
+	return nil
 }
 
 func openBlock(path, blockID string) (*tsdb.DBReadOnly, tsdb.BlockReader, error) {
@@ -213,34 +242,18 @@ func openBlock(path, blockID string) (*tsdb.DBReadOnly, tsdb.BlockReader, error)
 	if err != nil {
 		return nil, nil, err
 	}
-	blocks, err := db.Blocks()
+
+	if blockID == "" {
+		blockID, err = db.LastBlockID()
+		if err != nil {
+			return nil, nil, err
+		}
+	}
+
+	b, err := db.Block(blockID)
 	if err != nil {
 		return nil, nil, err
 	}
-	var block tsdb.BlockReader
-	if blockID != "" {
-		for _, b := range blocks {
-			if b.Meta().ULID.String() == blockID {
-				block = b
-				break
-			}
-		}
-	} else if len(blocks) > 0 {
-		block = blocks[len(blocks)-1]
-	}
-	if block == nil {
-		return nil, nil, fmt.Errorf("block %s not found", blockID)
-	}
-	return db, block, nil
-}
 
-// labelsToString returns labels as comma separated k=v pairs
-func labelsToString(lables map[string]string) string {
-	pairs := make([]string, len(lables))
-	i := 0
-	for k, v := range lables {
-		pairs[i] = fmt.Sprintf("%s=%s", k, v)
-		i++
-	}
-	return strings.Join(pairs, ", ")
+	return db, b, nil
 }
