@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
+	"github.com/thanos-io/objstore/client"
 	"math"
 	"math/rand"
 	"os"
@@ -11,20 +11,19 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-kit/kit/log"
-	"github.com/prometheus/client_golang/prometheus"
+	"github.com/go-kit/log"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 )
 
 func Test_e2e(t *testing.T) {
-	tmpDir := "./data"
+	tmpDir := "./tmp"
 	bktDir := tmpDir + "/bucket"
 	if err := os.MkdirAll(bktDir, 0750); err != nil {
-		t.Errorf("Unable to create %s directory for testing", bktDir)
+		t.Fatalf("Unable to create %s directory for testing", bktDir)
 	}
 	cacheDir := tmpDir + "/cache"
 	if err := os.MkdirAll(cacheDir, 0750); err != nil {
-		t.Errorf("Unable to create %s directory for testing", cacheDir)
+		t.Fatalf("Unable to create %s directory for testing", cacheDir)
 	}
 
 	// prepare test data
@@ -38,45 +37,45 @@ func Test_e2e(t *testing.T) {
 	f.Close()
 
 	// import metrics to local objStore
-	metrics := prometheus.NewRegistry()
 	logger := log.NewLogfmtLogger(log.NewSyncWriter(os.Stderr))
-	localStore := []byte(fmt.Sprintf(`{type: FILESYSTEM, config: {directory: %s}}`, bktDir))
-	blockSize := 2*time.Hour
+	bkt, err := client.NewBucket(logger, []byte("{type: FILESYSTEM, config: {directory: "+bktDir+"}}"), "thanos-kit")
+	if err != nil {
+		t.Fatalf("Open bucket: %v", err)
+	}
+	blockSize := 2 * time.Hour
 	importLabels := []string{
 		`prometheus="prometheus-a"`,
 		"datacenter=us",
 	}
-	if err := backfill(localStore, &inputFile, &blockSize, &cacheDir, &importLabels, logger, metrics); err != nil {
-		t.Errorf("Import of %s failed: %v", inputFile, err)
+	if err := importMetrics(bkt, &inputFile, &blockSize, &cacheDir, &importLabels, true, logger); err != nil {
+		t.Fatalf("Import of %s failed: %v", inputFile, err)
 	}
+	os.RemoveAll(cacheDir)
 
 	// check thanos labels
-	dirs, _ := ioutil.ReadDir(bktDir)
+	dirs, _ := os.ReadDir(bktDir)
 	ids := []string{}
 	for _, d := range dirs {
-		if d.Name() == "debug" {
-			continue
-		}
 		ids = append(ids, d.Name())
-		meta, err := metadata.Read(filepath.Join(bktDir, d.Name()))
+		meta, err := metadata.ReadFromDir(filepath.Join(bktDir, d.Name()))
 		if err != nil {
-			t.Errorf("fail to read meta.json for %s: %v", d.Name(), err)
+			t.Fatalf("fail to read meta.json for %s: %v", d.Name(), err)
 		}
 		if len(meta.Thanos.Labels) != 2 ||
 			meta.Thanos.Labels["prometheus"] != "prometheus-a" ||
 			meta.Thanos.Labels["datacenter"] != "us" {
-			t.Errorf("Wrong Thanos Labels in object storage block %s: %v", d.Name(), meta.Thanos.Labels)
+			t.Fatalf("Wrong Thanos Labels in object storage block %s: %v", d.Name(), meta.Thanos.Labels)
 		}
 	}
 
 	// export data from object storage
 	outFile := tmpDir + "/export.prom"
 	f, _ = os.Create(outFile)
-	metrics = prometheus.NewRegistry()
 	minT := int64(0)
 	maxT := int64(math.MaxInt64)
-	if err := dump(localStore, &ids, &cacheDir, &minT, &maxT, f, logger, metrics); err != nil {
-		t.Errorf("Export of %s failed: %v", ids, err)
+	match := "{__name__=~'(?s:.*)'}"
+	if err := dump(bkt, f, &ids, &cacheDir, &minT, &maxT, &match, logger); err != nil {
+		t.Fatalf("Export of %s failed: %v", ids, err)
 	}
 	f.Close()
 
@@ -84,7 +83,7 @@ func Test_e2e(t *testing.T) {
 	cmd := exec.Command("bash", "-c", fmt.Sprintf("diff <(sort %s) <(sort %s)", inputFile, outFile))
 	cmd.Stderr = os.Stderr
 	if err := cmd.Run(); err != nil {
-		t.Errorf("Compare files %s and %s result: %v", inputFile, outFile, err)
+		t.Fatalf("Compare files %s and %s result: %v", inputFile, outFile, err)
 	}
 
 	// Cleanup
